Crate rerun_bindings

The Rerun Python Log SDK.

This provides bindings between Python and Rust. It compiles into a Python wheel using https://github.com/PyO3/pyo3.

Feature flags

pypi — The features we turn on when building the rerun-sdk PyPi package for https://pypi.org/project/rerun-sdk/.

extension-module (enabled by default) — We need to enable the pyo3/extension-module when building the SDK, but we cannot enable it when building tests and benchmarks, so we must make it an opt-in feature.

https://pyo3.rs/latest/faq.html#i-cant-run-cargo-test-or-i-cant-build-in-a-cargo-workspace-im-having-linker-issues-like-symbol-not-found-or-undefined-reference-to-_pyexc_systemerror

https://pyo3.rs/latest/building_and_distribution.html#building-python-extension-modules

native_viewer (enabled by default) — Support spawning a native Rerun viewer with spawn().

web_viewer — Support serving a web viewer over HTTP with serve().

Enabling this adds quite a bit to the compile time and binary size, since it requires compiling and bundling the viewer as wasm.

You also need to install some additional tools, which you can do by running scripts/setup_web.sh.

Modules

arrow 🔒

Methods for handling Arrow datamodel log ingest

python_bridge 🔒

Statics

GLOBAL 🔒