rerun_bindings::catalog
Module entry

Structs

PyEntry

An entry in the catalog.

PyEntryId

A unique identifier for an entry in the catalog.

Enums

PyEntryKind

The kinds of entries that can be stored in the catalog.