Struct rerun_bindings::dataframe::PyComponentColumnDescriptor

struct PyComponentColumnDescriptor(ComponentColumnDescriptor);

The descriptor of a component column.

Component columns contain the data for a specific component of an entity.

Column descriptors are used to describe the columns in a [Schema][rerun.dataframe.Schema]. They are read-only. To select a component column, use [ComponentColumnSelector][rerun.dataframe.ComponentColumnSelector].

Tuple Fields
0: ComponentColumnDescriptor
Implementations

impl PyComponentColumnDescriptor

fn __repr__(&self) -> String

fn __eq__(&self, other: &Self) -> bool

fn entity_path(&self) -> String

The entity path.

This property is read-only.

fn component_name(&self) -> &str

The component name.

This property is read-only.

fn is_static(&self) -> bool

Whether the column is static.

This property is read-only.

impl PyComponentColumnDescriptor

unsafe fn __pymethod___eq____( py: Python<'_>, _raw_slf: *mut PyObject, arg0: *mut PyObject ) -> PyResult<*mut PyObject>

impl PyComponentColumnDescriptor

unsafe extern "C" fn __pymethod___richcmp____( slf: *mut PyObject, other: *mut PyObject, op: c_int ) -> *mut PyObject