rerun_bindings::dataframe
Struct PyRecording

pub struct PyRecording {
    pub(crate) store: ChunkStoreHandle,
    pub(crate) cache: QueryCacheHandle,
}

A single Rerun recording.

This can be loaded from an RRD file using [load_recording()][rerun.dataframe.load_recording].

A recording is a collection of data that was logged to Rerun. This data is organized as a column for each index (timeline) and each entity/component pair that was logged.

You can examine the [.schema()][rerun.dataframe.Recording.schema] of the recording to see what data is available, or create a [RecordingView][rerun.dataframe.RecordingView] to to retrieve the data.

Fields
store: ChunkStoreHandlecache: QueryCacheHandle
Implementations

impl PyRecording

fn engine(&self) -> QueryEngine<StorageEngine>

fn find_best_component( &self, entity_path: &EntityPath, component_name: &str, ) -> ComponentName

fn extract_contents_expr( &self, expr: Bound<'_, PyAny>, ) -> PyResult<ViewContentsSelector>

Convert a ViewContentsLike into a ViewContentsSelector.

ViewContentsLike = Union[str, Dict[str, Union[ComponentLike, Sequence[ComponentLike]]]]

We cant do this with the normal FromPyObject mechanisms because we want access to the QueryEngine to resolve the entity paths.

impl PyRecording

fn schema(&self) -> PySchema

The schema describing all the columns available in the recording.

fn view( slf: Bound<'_, Self>, index: &str, contents: Bound<'_, PyAny>, include_semantically_empty_columns: bool, include_indicator_columns: bool, include_tombstone_columns: bool, ) -> PyResult<PyRecordingView>

Create a [RecordingView][rerun.dataframe.RecordingView] of the recording according to a particular index and content specification.

The only type of index currently supported is the name of a timeline.

The view will only contain a single row for each unique value of the index that is associated with a component column that was included in the view. Component columns that are not included via the view contents will not impact the rows that make up the view. If the same entity / component pair was logged to a given index multiple times, only the most recent row will be included in the view, as determined by the row_id column. This will generally be the last value logged, as row_ids are guaranteed to be monotonically increasing when data is sent from a single process.

Parameters

index : str The index to use for the view. This is typically a timeline name. contents : ViewContentsLike The content specification for the view.

This can be a single string content-expression such as: `"world/cameras/**"`, or a dictionary
specifying multiple content-expressions and a respective list of components to select within
that expression such as `{"world/cameras/**": ["ImageBuffer", "PinholeProjection"]}`.

include_semantically_empty_columns : bool, optional Whether to include columns that are semantically empty, by default False.

Semantically empty columns are components that are `null` or empty `[]` for every row in the recording.

include_indicator_columns : bool, optional Whether to include indicator columns, by default False.

Indicator columns are components used to represent the presence of an archetype within an entity.

include_tombstone_columns : bool, optional Whether to include tombstone columns, by default False.

Tombstone columns are components used to represent clears. However, even without the clear
tombstone columns, the view will still apply the clear semantics when resolving row contents.

Returns

RecordingView The view of the recording.

Examples

All the data in the recording on the timeline โmy_indexโ:

recording.view(index="my_index", contents="/**")

Just the Position3D components in the โpointsโ entity:

recording.view(index="my_index", contents={"points": "Position3D"})

fn recording_id(&self) -> String

The recording ID of the recording.

fn application_id(&self) -> PyResult<String>

The application ID of the recording.