rerun_bindings::python_bridge
Struct PyRecordingStream

struct PyRecordingStream(RecordingStream);

Tuple Fields
0: RecordingStream
Implementations

impl PyRecordingStream

fn is_forked_child(&self) -> bool

Determine if this stream is operating in the context of a forked child process.

This means the stream was created in the parent process. It now exists in the child process by way of fork, but it is effectively a zombie since its batcher and sink threads would not have been copied.

Calling operations such as flush or set_sink will result in an error.