Struct rerun_bindings::python_bridge::ViewCoordinates

#[repr(transparent)]
pub struct ViewCoordinates(pub [u8; 3]);

Component: How we interpret the coordinate system of an entity/space.

For instance: What is “up”? What does the Z axis mean? Is this right-handed or left-handed?

The three coordinates are always ordered as [x, y, z].

For example [Right, Down, Forward] means that the X axis points to the right, the Y axis points down, and the Z axis points forward.

The following constants are used to represent the different directions. Up = 1 Down = 2 Right = 3 Left = 4 Forward = 5 Back = 6

Tuple Fields
0: [u8; 3]

The directions of the [x, y, z] axes.

Implementations

impl ViewCoordinates

pub const fn new(x: ViewDir, y: ViewDir, z: ViewDir) -> ViewCoordinates

Construct a new ViewCoordinates from an array of ViewDirs.

pub fn from_up_and_handedness( up: SignedAxis3, handedness: Handedness ) -> ViewCoordinates

Choses a coordinate system based on just an up-axis.

pub fn sanity_check(&self) -> Result<(), String>

Returns an error if this does not span all three dimensions.

pub fn up(&self) -> Option<SignedAxis3>

pub fn right(&self) -> Option<SignedAxis3>

pub fn forward(&self) -> Option<SignedAxis3>

pub fn describe_short(&self) -> String

pub fn describe(&self) -> String

pub fn from_other(&self, other: &ViewCoordinates) -> Mat3

Returns a matrix that transforms from another coordinate system to this (self) one.

pub fn to_rdf(&self) -> Mat3

Returns a matrix that transforms this coordinate system to RDF.

(RDF: X=Right, Y=Down, Z=Forward)

pub fn from_rdf(&self) -> Mat3

Returns a matrix that transforms from RDF to this coordinate system.

(RDF: X=Right, Y=Down, Z=Forward)

pub fn to_rub(&self) -> Mat3

Returns a matrix that transforms this coordinate system to RUB.

(RUB: X=Right, Y=Up, Z=Back)

pub fn from_rub(&self) -> Mat3

Returns a matrix that transforms from RUB to this coordinate system.

(RUB: X=Right, Y=Up, Z=Back)

pub fn from_rub_quat(&self) -> Result<Quat, String>

Returns a quaternion that rotates from RUB to this coordinate system.

Errors if the coordinate system is left-handed or degenerate.

(RUB: X=Right, Y=Up, Z=Back)

pub fn handedness(&self) -> Result<Handedness, String>

impl ViewCoordinates

pub const ULF: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Left as u8, ViewDir::Forward as u8])

pub const UFL: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Forward as u8, ViewDir::Left as u8])

pub const LUF: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Up as u8, ViewDir::Forward as u8])

pub const LFU: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Forward as u8, ViewDir::Up as u8])

pub const FUL: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Up as u8, ViewDir::Left as u8])

pub const FLU: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Left as u8, ViewDir::Up as u8])

pub const ULB: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Left as u8, ViewDir::Back as u8])

pub const UBL: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Back as u8, ViewDir::Left as u8])

pub const LUB: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Up as u8, ViewDir::Back as u8])

pub const LBU: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Back as u8, ViewDir::Up as u8])

pub const BUL: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Up as u8, ViewDir::Left as u8])

pub const BLU: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Left as u8, ViewDir::Up as u8])

pub const URF: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Right as u8, ViewDir::Forward as u8])

pub const UFR: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Forward as u8, ViewDir::Right as u8])

pub const RUF: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Up as u8, ViewDir::Forward as u8])

pub const RFU: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Forward as u8, ViewDir::Up as u8])

pub const FUR: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Up as u8, ViewDir::Right as u8])

pub const FRU: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Right as u8, ViewDir::Up as u8])

pub const URB: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Right as u8, ViewDir::Back as u8])

pub const UBR: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Back as u8, ViewDir::Right as u8])

pub const RUB: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Up as u8, ViewDir::Back as u8])

pub const RBU: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Back as u8, ViewDir::Up as u8])

pub const BUR: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Up as u8, ViewDir::Right as u8])

pub const BRU: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Right as u8, ViewDir::Up as u8])

pub const DLF: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Left as u8, ViewDir::Forward as u8])

pub const DFL: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Forward as u8, ViewDir::Left as u8])

pub const LDF: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Down as u8, ViewDir::Forward as u8])

pub const LFD: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Forward as u8, ViewDir::Down as u8])

pub const FDL: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Down as u8, ViewDir::Left as u8])

pub const FLD: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Left as u8, ViewDir::Down as u8])

pub const DLB: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Left as u8, ViewDir::Back as u8])

pub const DBL: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Back as u8, ViewDir::Left as u8])

pub const LDB: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Down as u8, ViewDir::Back as u8])

pub const LBD: ViewCoordinates = Self([ViewDir::Left as u8, ViewDir::Back as u8, ViewDir::Down as u8])

pub const BDL: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Down as u8, ViewDir::Left as u8])

pub const BLD: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Left as u8, ViewDir::Down as u8])

pub const DRF: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Right as u8, ViewDir::Forward as u8])

pub const DFR: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Forward as u8, ViewDir::Right as u8])

pub const RDF: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Down as u8, ViewDir::Forward as u8])

pub const RFD: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Forward as u8, ViewDir::Down as u8])

pub const FDR: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Down as u8, ViewDir::Right as u8])

pub const FRD: ViewCoordinates = Self([ViewDir::Forward as u8, ViewDir::Right as u8, ViewDir::Down as u8])

pub const DRB: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Right as u8, ViewDir::Back as u8])

pub const DBR: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Back as u8, ViewDir::Right as u8])

pub const RDB: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Down as u8, ViewDir::Back as u8])

pub const RBD: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Back as u8, ViewDir::Down as u8])

pub const BDR: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Down as u8, ViewDir::Right as u8])

pub const BRD: ViewCoordinates = Self([ViewDir::Back as u8, ViewDir::Right as u8, ViewDir::Down as u8])

pub const RIGHT_HAND_X_UP: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Right as u8, ViewDir::Forward as u8])

pub const RIGHT_HAND_X_DOWN: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Right as u8, ViewDir::Back as u8])

pub const RIGHT_HAND_Y_UP: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Up as u8, ViewDir::Back as u8])

pub const RIGHT_HAND_Y_DOWN: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Down as u8, ViewDir::Forward as u8])

pub const RIGHT_HAND_Z_UP: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Forward as u8, ViewDir::Up as u8])

pub const RIGHT_HAND_Z_DOWN: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Back as u8, ViewDir::Down as u8])

pub const LEFT_HAND_X_UP: ViewCoordinates = Self([ViewDir::Up as u8, ViewDir::Right as u8, ViewDir::Back as u8])

pub const LEFT_HAND_X_DOWN: ViewCoordinates = Self([ViewDir::Down as u8, ViewDir::Right as u8, ViewDir::Forward as u8])

pub const LEFT_HAND_Y_UP: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Up as u8, ViewDir::Forward as u8])

pub const LEFT_HAND_Y_DOWN: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Down as u8, ViewDir::Back as u8])

pub const LEFT_HAND_Z_UP: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Back as u8, ViewDir::Up as u8])

pub const LEFT_HAND_Z_DOWN: ViewCoordinates = Self([ViewDir::Right as u8, ViewDir::Forward as u8, ViewDir::Down as u8])