rerun_bindings::video
Function asset_video_read_frame_timestamps_nanos

pub fn asset_video_read_frame_timestamps_nanos(
    video_bytes_arrow_array: &Bound<'_, PyAny>,
    media_type: Option<&str>,
) -> PyResult<Vec<i64>>

Reads the timestamps of all frames in a video asset.

Implementation note: On the Python side we start out with a pyarrow array of bytes. Converting it to Python bytes can be done with to_pybytes but this requires copying the data. So instead, we pass the arrow array directly.