rerun_bindings
Module dataframe

Structs

ComponentLike

PyComponentColumnDescriptor

The descriptor of a component column.

PyComponentColumnSelector

A selector for a component column.

PyDataFusionTable

PyIndexColumnDescriptor 🔒

The descriptor of an index column.

PyIndexColumnSelector

A selector for an index column.

PyRRDArchive

An archive loaded from an RRD.

PyRecording

A single Rerun recording.

PyRecordingView

A view of a recording restricted to a given index, containing a specific set of entities and components.

PySchema

SchemaIterator

Enums

AnyColumn 🔒

A type alias for any component-column-like object.

AnyComponentColumn

A type alias for any component-column-like object.

IndexValuesLike 🔒

A type alias for index values.

PyRecordingHandle

Functions

__pyfunction_load_archive 🔒 ⚠

__pyfunction_load_recording 🔒 ⚠

load_archive

Load a rerun archive from an RRD file.

load_recording

Load a single recording from an RRD file.

py_rerun_warn 🔒

register 🔒

Register the rerun.dataframe module.